[CHANGE] cc margin to float

+++ custom_cctz/models/crm_inherit_crm_lead.py
@@ -12,7 +12,7 @@ class CrmLead(models.Model):
     purchase_time_frame_id = fields.Many2one('purchase.time', string="Purchase Time Frame")
     payment_terms_id = fields.Many2one('payment.terms', string="Payment Terms")
     status_id = fields.Many2one('crm.status', string="status")
-    cc_margin = fields.Integer(string="CC Margin")
+    cc_margin = fields.Float(string="CC Margin")
     tender_status = fields.Selection([
         ('Yes', 'Yes'),
         ('No', 'No')
